backend/requirements.txt
requests
beautifulsoup4
tqdm
numpy
pandas
tabulate
glob2
sentence-transformers
faiss-cpu
scikit-learn
torch
transformers
python-dotenv
Flask
flask-cors
groq
aiohttp
fuzzywuzzy
python-Levenshtein

